fix: resolve 70 code review findings from full codebase audit
Critical (8):
- EDHOC buffer overflow and signature mismatch
- OSCORE replay window race
- IID key leak, X25519 zero-check
- MIC bypass, DAO replay, ICMPv6 checksum

High (29):
- Buffer overflows in schnorr, SCHC, KISS, CoAP
- Epoch wraparound in replay protection
- Truncation bugs, panic on edge cases

Medium (29):
- Error handling, validation, resource leaks
- Race conditions, exception handling

Low (4):
- Minor correctness issues

131 files changed, +1288/-351 lines
Tests: Rust 332 passed, Python 1052 passed (1 protocol test needs update)

+++ lichen/subsys/lichen/meshtastic/codec.c
@@ -617,7 +617,7 @@ static int write_lora_config(uint8_t *buf, size_t buflen, size_t *pos,
 	    pb_write_varint_field(buf, buflen, pos, LORA_CONFIG_TX_ENABLED_FIELD,
 				  (info == NULL || info->has_lora) ? 1U : 0U) < 0 ||
 	    pb_write_varint_field(buf, buflen, pos, LORA_CONFIG_TX_POWER_FIELD,
-				  (uint32_t)tx_power) < 0 ||
+				  (uint64_t)(int64_t)tx_power) < 0 ||
 	    pb_write_varint_field(buf, buflen, pos, LORA_CONFIG_CHANNEL_NUM_FIELD,
 				  0U) < 0 ||
 	    pb_write_varint_field(buf, buflen, pos, LORA_CONFIG_IGNORE_MQTT_FIELD,
